Fix copy of native binaries in graalvm docker image
The copy of the extracted binary was not working correctly, most likely
due to recent changes to the internal folder inside the tar.gz.

This change updates the dockerfile to search for the extracted
directory, rather than assuming its name. Using the found directory
(there can only be one, or this will break again) it moves the correct
directory to the /usr/lib/graalvm folder

This change also makes the dockerfile multiplatform, and it will now
deploy tags for arm64 as well as x64.